Move integration tests out of repo root
