+++ tests/integration/e2e_test.go
@@ -1,4 +1,4 @@
-package main_test
+package integration_test
 
 import (
 	"bytes"
