+++ tests/integration/route_backend_pool_test.go
@@ -1,4 +1,4 @@
-package main_test
+package integration_test
 
 import (
 	"context"
