+++ tests/integration/traffic_tracer_test.go
@@ -1,4 +1,4 @@
-package main_test
+package integration_test
 
 import (
 	"context"
